Make some optimizations and change status() method

+++ module/routes.py
@@ -9,10 +9,6 @@ users = orders.users
 partial_content = 'You must provide the required values.'
 
 
-def get_json(key):
-    return request.get_json()[key]
-
-
 @app.route('/')
 def welcome():
     return "<h1>Welcome to the Fast-Food-Fast API<h1>"
@@ -22,10 +18,10 @@ def welcome():
 def register():
     "This route registers a new user."
     try:
-        name = get_json('username')
-        email = get_json('email')
-        location = get_json('location')
-        key_point = get_json('key point')
+        name = request.get_json()['username']
+        email = request.get_json()['email']
+        location = request.get_json()['location']
+        key_point = request.get_json()['key point']
     except KeyError:
         return jsonify(partial_content + '[username, email, location,\
  key point]'), 400
@@ -36,53 +32,48 @@ def register():
     return response
 
 
-@app.route('/api/v1/orders', methods=['POST'])
+@app.route('/api/v1/orders', methods=['GET', 'POST'])
 def place_order():
     "This route adds a new order to the orders list when parsed with a json\
  object containing all required values."
-    try:
-        name = get_json('name')
-        quantity = get_json('quantity')
-        comment = get_json('comment')
-        customer = get_json('username')
-    except KeyError:
-        return jsonify(partial_content + ' [name, quantity, comment,\
+    if request.method == 'POST':
+        try:
+            name = request.get_json()['name']
+            quantity = request.get_json()['quantity']
+            comment = request.get_json()['comment']
+            customer = request.get_json()['username']
+        except KeyError:
+            return jsonify(partial_content + ' [name, quantity, comment,\
  username]'), 400
 
-    order = Order(name, quantity, comment, customer)
-    response = orders.place_order(order)
+        order = Order(name, quantity, comment, customer)
+        response = orders.place_order(order)
+        return response
 
-    return response
-
-
-@app.route('/api/v1/orders', methods=['GET'])
-def all_orders():
-    "This route returns the list of orders."
-    response = orders.all_orders()
-    return response
+    elif request.method == 'GET':
+        "This route returns the list of orders."
+        response = orders.all_orders()
+        return response
 
 
-@app.route('/api/v1/orders/<int:order_id>', methods=['GET'])
+@app.route('/api/v1/orders/<int:order_id>', methods=['GET', 'PUT'])
 def get_order(order_id):
     "This route returns the details of a particular order."
-    order = order_id
-    response = orders.get_order(order)
+    if request.method == 'GET':
+        order = order_id
+        response = orders.get_order(order)
+        return response
 
-    return response
+    elif request.method == 'PUT':
+        "This route updates the status key of a particular order."
+        order = order_id
+        try:
+            state = request.get_json()['status']
+        except KeyError:
+            return jsonify(partial_content + ' [status]'), 400
 
-
-@app.route('/api/v1/orders/<int:order_id>', methods=['PUT'])
-def update_order(order_id):
-    "This route updates the status key of a particular order."
-    order = order_id
-    try:
-        state = get_json('status')
-    except KeyError:
-        return jsonify(partial_content + ' [status]'), 400
-
-    response = orders.update_order(order, state)
-
-    return response
+        response = orders.update_order(order, state)
+        return response
 
 
 @app.route('/api/v1/menu', methods=['GET'])
@@ -96,12 +87,12 @@ def menu():
 def add_food_item():
     "This route adds a food item to the food list."
     try:
-        name = get_json('name')
-        price = get_json('price')
-        ready_in = get_json('ready in')
-        status = get_json('status')
-        units = get_json('units')
-        tags = get_json('tags')
+        name = request.get_json()['name']
+        price = request.get_json()['price']
+        ready_in = request.get_json()['ready in']
+        status = request.get_json()['status']
+        units = request.get_json()['units']
+        tags = request.get_json()['tags']
     except KeyError:
         return jsonify(partial_content + ' [name, price, ready in,\
  status, units, tags'), 400
@@ -114,18 +105,17 @@ def add_food_item():
     return response
 
 
-@app.route('/api/v1/menu/<string:name>', methods=['PUT'])
+@app.route('/api/v1/menu/<string:name>', methods=['PUT', 'DELETE'])
 def update_food_item(name):
-    "This route updates the details of a particular food item."
-    food_name = name
-    updates = request.get_json()
-    response = foods.update_food_item(food_name, updates)
-    return response
+    if request.method == 'PUT':
+        "This route updates the details of a particular food item."
+        food_name = name
+        updates = request.get_json()
+        response = foods.update_food_item(food_name, updates)
+        return response
 
-
-@app.route('/api/v1/menu/<string:name>', methods=['DELETE'])
-def delete_food_item(name):
-    "This route deletes a particular food item from the food list."
-    food_name = name
-    response = foods.delete_food_item(food_name)
-    return response
+    elif request.method == 'DELETE':
+        "This route deletes a particular food item from the food list."
+        food_name = name
+        response = foods.delete_food_item(food_name)
+        return response
